Work on console.
